\name{datasets}
\alias{timage12}
\alias{crimerate}
\title{data used for demo in SVDgen}
\description{The \code{crimerate} dataset provides crime rates per 100,000 people in
 seven categories for each of the fifty states in 1977.  The \code{timage12} dataset
  is an image from fMRI analysis, it is a \emph{t}-statistic image over 12 subjects of the activation (verbal) parameter. }
\usage{
data(crimerate)
data(timage12)
}
\format{
A matrix of \code{50 x 7} for the \code{crimerate} data.

A matrix \code{91 x 109} for \code{timage12} data.}
\source{\code{crimerate} comes from SAS. The image comes from FMRIB center, University of Oxford. }

\keyword{datasets}
